requirements/test.txt
factory_boy==3.3.0
pytest==7.4.0
pytest-django==4.5.2
